+++ script/provider/diagnostic.lua
@@ -183,8 +183,7 @@ function m.doDiagnostic(uri)
         return
     end
     uri = files.asKey(uri)
-    if (files.isLibrary(uri) or ws.isIgnored(uri))
-    and not files.isOpen(uri) then
+    if files.isLibrary(uri) or ws.isIgnored(uri) then
         return
     end