+++ changelog.md
@@ -1,7 +1,7 @@
 # changelog
 
 ## 2.1.0
-* `NEW` supports local config file, using `--configpath="config.json"`
+* `NEW` supports local config file, using `--configpath="config.json"`, [learn more here](https://github.com/sumneko/lua-language-server/wiki/Setting-without-VSCode)
 
 ## 2.0.5
 `2021-7-1`