+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 2.1.0
 * `NEW` supports local config file, using `--configpath="config.json"`, [learn more here](https://github.com/sumneko/lua-language-server/wiki/Setting-without-VSCode)
+* `NEW` goto `type definition`
 
 ## 2.0.5
 `2021-7-1`