+++ changelog.md
@@ -1,6 +1,7 @@
 # changelog
 
 ## 2.1.0
+`2021-7-2`
 * `NEW` supports local config file, using `--configpath="config.json"`, [learn more here](https://github.com/sumneko/lua-language-server/wiki/Setting-without-VSCode)
 * `NEW` goto `type definition`
 * `NEW` infer type by callback param: